requirements.txt
ffmpeg-python==0.2.0
future==1.0.0
pathlib==1.0.1
PyQt6==6.10.1
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.3
python-vlc==3.0.21203
